fix test failure

+++ src/groupnorm.cpp
@@ -47,7 +47,7 @@ miopenStatus_t GroupNormForward(Handle& handle,
                                 const TensorDescriptor& rstdDesc,
                                 Data_t rstd,
                                 miopenNormMode_t mode,
-                                int32_t num_groups,
+                                uint64_t num_groups,
                                 float epsilon)
 {
     const auto problem = groupnorm::ProblemDescription{
